Pickups
Speed up and speed down

+++ Assets/Scripts/GravityChange.cs
@@ -6,7 +6,7 @@ public class GravityChange : MonoBehaviour
 {
 
     public bool reverseGravity = false;
-    public int gravitySpeed = 7;
+    public float gravitySpeed = 7;
 
     // Start is called before the first frame update
     void Start()
